[klas3bis] Move kernelAccess and indexingFunction functions to a separate MemoryAccess.h - all ok
This will need a much more throught cleanup, once we agree better what is needed in the Fortran bridge

But for the moment this is a reasonable separation of data access and computations

On itscrd70.cern.ch [CPU: Intel(R) Xeon(R) Silver 4216 CPU] [GPU: 1x Tesla V100S-PCIE-32GB]:
=========================================================================
Process                     = EPOCH1_EEMUMU_CUDA [nvcc 11.1.105 (gcc 10.2.0)] [inlineHel=0] [hardcodeCIPC=0]
FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
EvtsPerSec[MatrixElems] (3) = ( 5.525339e+08                 )  sec^-1
EvtsPerSec[MECalcOnly] (3a) = ( 1.162881e+09                 )  sec^-1
MeanMatrixElemValue         = ( 1.371706e-02 +- 3.270315e-06 )  GeV^0
TOTAL       :     1.123886 sec
       785,576,398      cycles:u                  #    0.880 GHz
     1,490,945,195      instructions:u            #    1.90  insn per cycle
       1.427680604 seconds time elapsed
==PROF== Profiling "sigmaKin": launch__registers_per_thread 130
==PROF== Profiling "sigmaKin": sm__sass_average_branch_targets_threads_uniform.pct 100%
=========================================================================
Process                     = EPOCH1_EEMUMU_CPP [gcc 10.2.0] [inlineHel=0] [hardcodeCIPC=0]
FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
Internal loops fptype_sv    = SCALAR ('none': ~vector[1], no SIMD)
OMP threads / `nproc --all` = 1 / 4
EvtsPerSec[MECalcOnly] (3a) = ( 1.301289e+06                 )  sec^-1
MeanMatrixElemValue         = ( 1.371706e-02 +- 3.270315e-06 )  GeV^0
TOTAL       :     7.290284 sec
    19,162,611,591      cycles:u                  #    2.621 GHz
    48,318,763,032      instructions:u            #    2.52  insn per cycle
       7.316508951 seconds time elapsed
=Symbols in CPPProcess.o= (~sse4:  576) (avx2:    0) (512y:    0) (512z:    0)
-------------------------------------------------------------------------
Process                     = EPOCH1_EEMUMU_CPP [gcc 10.2.0] [inlineHel=0] [hardcodeCIPC=0]
FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
Internal loops fptype_sv    = VECTOR[4] ('512y': AVX512, 256bit) [cxtype_ref=YES]
OMP threads / `nproc --all` = 1 / 4
EvtsPerSec[MECalcOnly] (3a) = ( 4.851138e+06                 )  sec^-1
MeanMatrixElemValue         = ( 1.371706e-02 +- 3.270315e-06 )  GeV^0
TOTAL       :     3.716292 sec
     9,109,552,318      cycles:u                  #    2.436 GHz
    16,273,698,120      instructions:u            #    1.79  insn per cycle
       3.743027573 seconds time elapsed
=Symbols in CPPProcess.o= (~sse4:    0) (avx2: 2762) (512y:   52) (512z:    0)
=========================================================================
